unittesting and bugfixing for: * UnsprungAccelerometerSystem * Settings

+++ AutomaticSuspensionSystem/src/FrequencySystem.h
@@ -38,7 +38,7 @@ public:
 		return processing;
 	}
 
-	unsigned short getAverageTime() {
+	unsigned long getAverageTime() {
 		return timing.getAverage();
 	}
